Re-initialize the whole UVC stack on error
Description:
Still seeing intermittent "Access" failure to acquire the camera.

One potential cause is when the camera was active at the time the program was killed and restarted, which does reproduce the problem. Unfortunately there doesn't seem to be a solution, short of killing and re-running the program again.

This diff adds more opportunities to retry getting the camera, especially to repeatedly retry with a fresh context.

At least after this, unplugging the camera and plugging it back in seems to fix the problem. In the future when this is running under Docker with restarts, we might actually opt to crash and have the process restarted, but that is for another day.

Test Plan:
Kill and launch the program a bunch of times (trying to do it when the camera activity light is lit).
Still doesn't recover, unfortunately.