Schedule recording changed frames
Description:
On a schedule (weekdays at 7am), turn on the camera and watch for changes to the video, using the Hamming distance between consecutive frames to tell if something is happening in view.

When there has been a chance, save the frame to disk. Otherwise let it pass.

Test Plan:
Add a command-line option to enable the camera immediately, e.g.
```
cargo run ~/Downloads/porch/images --start-watching-immediately
```