Calculate inference image on capture
Description:
The PyTorch model is apparently trained on images in the form of RGB f32s (range 0.0 - 1.0). The images are scaled to dimensions of 224x224.

In order to do inferencing on device, we need our camera input to match this format.

This diff uses opencv to transform incoming frames into this format, to be used for inference in the next diff.

Test Plan:
Write out the resized image as a jpeg as a debugging mechanism.