Tracing for frame processing
Description:
Inference is expensive, so let's instrument how long capture takes before adding inference to the process.

Instrumenting async code is potentially subtle. Comments on tracing APIs warn about using spans across async dispatches. They seem to assume that we only want to know the inclusive time, not the wall time. I want to know both, so I use both spans (which do not suspend the trace) and the `#[tracing::instrument` macro (which does suspend the trace).

Test Plan:
Changed the `--start-watching-immediately` flag to trace for one minute only, so that the capture will complete and the trace file be saved to disk.
Run it and confirm that the trace is saved and looks correct.