Fix various problems with tensor calculation
Description:
We need to make the image tensor match the color space and tensor shape that the PyTorch PIL process uses. Since we are using OpenCV on the device rather than Python, there are several hoops we need to jump through. This solution was arrived at via several rounds of trial and error and research.

- Convert from BGR to RGB before calculating tensor. An earlier commit removed this step in error.
- Use libtorch (via tch crate) to transpose the matrix when converting it to a vector.

Test Plan:
Tested end to end with the rest of the stack.
Inference outputs seem to work as expected now, able to correctly detect if the gate is open or closed for real input.